frontend/manageusers/requirements.txt
boto3>=1.21.33
awscli>=1.20.0
grpcio
grpcio-tools